requirements.txt
Django==1.11
django-cors-headers==2.1.0 # 解决跨域访问问题
djangorestframework==3.6.3
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.0
django-guardian==1.4.9
django-import-export==1.0.0
django-qiniu-storage==2.3.1
django-redis==4.8.0
django-reversion==2.0.9
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
mysqlclient
celery==4.0.2
Pillow==4.0.0
future==0.16.0
coreapi==2.3.1